move:결제 관련 로직을 products로 이동
