+++ popcorngeek/urls.py
@@ -27,6 +27,7 @@ urlpatterns = [
     path("api/v1/accounts/", include("accounts.urls")),
     path("api/v1/movies/", include("movies.urls")),
     path("api/v1/reviews/", include("reviews.urls")),
+    path("api/v1/priducts/", include("priducts.urls")),
 ]
 
 if settings.DEBUG:
